apps/api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.23
asyncpg==0.28.0
aiosqlite==0.19.0
apscheduler==3.10.4
aiohttp==3.8.6
supabase>=2.0.3
python-dotenv==1.0.0
httpx==0.25.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
alembic==1.12.0
# Removed gotrue==1.1.0 as it conflicts with supabase which requires gotrue>=1.3.0
# Removed psycopg2-binary as it conflicts with asyncpg for async SQLAlchemy
